Improved documentation and bug fixing
1. EpiCSeg fails before segmenting when the provided regions
comprise only a single bin per region. I also added a test for that
and improved the command line documentation.

2. improved documentation of the normalizecounts subprogram, now
it is clearly stated what type of matrices can be normalized together.

3. found and fixed a completely unrelated bug in normalizecounts that is
was triggered when using multiple threads. I am not sure why I didn't
see it earlier (it was triggered by the standard unit tests). It
was related to a stupid initialization "int a, b = 3" which leaves
"a" undefined.